Support terraform PRs from non-admins (#362)
* Support terraform PRs from non-admins.

The terraform formatting action was breaking when a non-admin
created a PR. This revises that and makes the token usages
more consistent across the plan and apply files.

* Remove the auto-committing of formatting changes.

This complicates things when outside contributors suggest changes.
It's also a potential security vulnerability to run an outside
contributor's code with elevated permissions. This now will cause
any formatting issues to fail the github action and will need to
be fixed manually.

* Restore the variable param for terraform jobs.

The organization owner block requires the token to be specified:
https://registry.terraform.io/providers/integrations/github/latest/docs\#argument-reference

* Reflect django-valkey's discussions change.

---------

Co-authored-by: Daniel M <cunla@users.noreply.github.com>