Create an super admins team and switch admins to moderators (#305)
* Create an super admins team and switch admins to moderators

The super admins team will have admin permissions in the GitHub org
while the admins team will have moderator permissions.

This needs to be followed up with another commit to reduce the
permissions of the org admins teams to member. This is being
split up to avoid removing all admin permissions.

* Refine the descriptions of admin teams.